feat(project): add validation so that only the manager can update or delete the project

+++ server/src/controllers/ProjectController.ts
@@ -52,6 +52,9 @@ export class ProjectController {
             if (!project) {
                 return res.status(404).json({error: "Project not found"})
             }
+            if (project.manager.toString() !== req.user.id.toString()) {
+                return res.status(404).json({error: "Solo el manager puede actualizar el proyecto"})
+            }
             project.clientName = req.body.clientName
             project.projectName = req.body.projectName
             project.description = req.body.description
@@ -68,6 +71,9 @@ export class ProjectController {
             const project = await Project.findById(id)
             if (!project) {
                 return res.status(404).json({error: "Project not found"})
+            }
+            if (project.manager.toString() !== req.user.id.toString()) {
+                return res.status(404).json({error: "Solo el manager puede eliminar el proyecto"})
             }
             await project.deleteOne()
             res.send("Project deleted")
